+++ GifExceptionBundle.php
@@ -19,7 +19,5 @@ class GifExceptionBundle extends Bundle
     public function build(ContainerBuilder $container)
     {
         parent::build($container);
-
-        $container->addCompilerPass($this->getContainerExtension());
     }
 }